diorama-choose/requirements.txt
pyserial==3.4
cobs>=1.1.3,<1.2
